+++ packages/canonical-bridge-widget/src/modules/aggregator/providers/TokenPricesProvider.tsx
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { useQuery } from '@tanstack/react-query';
 import { useCallback, useEffect } from 'react';
-import { IBridgeToken } from '@bnb-chain/canonical-bridge-sdk';
+import { ChainType, isNativeToken } from '@bnb-chain/canonical-bridge-sdk';
 
 import { useBridgeConfig } from '@/CanonicalBridgeProvider';
 import { TIME } from '@/core/constants';
@@ -19,29 +19,14 @@ interface ITokenPricesResponse {
 }
 
 export function TokenPricesProvider() {
-  const bridgeConfig = useBridgeConfig();
   const dispatch = useAppDispatch();
+  const { fetchApiTokenPrices } = useTokenPrice();
 
   const { isLoading, data } = useQuery<TokenPricesContextProps>({
     staleTime: TIME.MINUTE * 5,
     refetchInterval: TIME.MINUTE * 5,
     queryKey: ['tokenPrices'],
-    queryFn: async () => {
-      const { serverEndpoint } = bridgeConfig.http;
-
-      const [cmcRes, llamaRes] = await Promise.allSettled([
-        axios.get<ITokenPricesResponse>(`${serverEndpoint}/api/token/cmc`),
-        axios.get<ITokenPricesResponse>(`${serverEndpoint}/api/token/llama`),
-      ]);
-
-      const cmcPrices = cmcRes.status === 'fulfilled' ? cmcRes.value.data.data : {};
-      const llamaPrices = llamaRes.status === 'fulfilled' ? llamaRes.value.data.data : {};
-
-      return {
-        cmcPrices,
-        llamaPrices,
-      };
-    },
+    queryFn: async () => fetchApiTokenPrices(),
   });
 
   useEffect(() => {
@@ -54,27 +39,32 @@ export function TokenPricesProvider() {
 
 export function useTokenPrice() {
   const tokenPrices = useAppSelector((state) => state.aggregator.tokenPrices);
+  const bridgeConfig = useBridgeConfig();
+
+  const { serverEndpoint } = bridgeConfig.http;
 
   const getTokenPrice = useCallback(
-    (params: IBridgeToken | { symbol?: string; address?: string } = {}) => {
-      const tokenSymbol = (params as IBridgeToken).displaySymbol;
-      const tokenAddress = params.address;
-
-      if (tokenSymbol && tokenAddress) {
+    ({
+      chainId,
+      chainType,
+      tokenAddress,
+    }: { chainId?: number; chainType?: ChainType; tokenAddress?: string } = {}) => {
+      if (chainId && chainType && tokenAddress) {
         const { cmcPrices, llamaPrices } = tokenPrices;
 
-        const key1 = `${tokenSymbol?.toLowerCase()}:${tokenAddress?.toLowerCase()}`;
-        const key3 = tokenSymbol?.toLowerCase();
-        const key2 = `ethereum:${key3}`;
+        let key = '';
+        const isNative = isNativeToken(tokenAddress, chainType);
+        if (isNative) {
+          key = `${chainId}`;
+        } else {
+          if (chainType === 'evm') {
+            key = `${chainId}:${tokenAddress.toLowerCase()}`;
+          } else {
+            key = `${chainId}:${tokenAddress}`;
+          }
+        }
 
-        let price =
-          cmcPrices?.[key1]?.price ??
-          llamaPrices?.[key1]?.price ??
-          cmcPrices?.[key2]?.price ??
-          llamaPrices?.[key2]?.price ??
-          cmcPrices?.[key3]?.price ??
-          llamaPrices?.[key3]?.price;
-
+        let price = cmcPrices?.[key]?.price ?? llamaPrices[key]?.price;
         if (price !== undefined) {
           price = Number(price);
         }
@@ -85,7 +75,45 @@ export function useTokenPrice() {
     [tokenPrices],
   );
 
+  const fetchApiTokenPrices = useCallback(async () => {
+    const [cmcRes, llamaRes] = await Promise.allSettled([
+      axios.get<ITokenPricesResponse>(`${serverEndpoint}/api/token/v2/cmc`),
+      axios.get<ITokenPricesResponse>(`${serverEndpoint}/api/token/v2/llama`),
+    ]);
+
+    const cmcPrices = cmcRes.status === 'fulfilled' ? cmcRes.value.data.data : {};
+    const llamaPrices = llamaRes.status === 'fulfilled' ? llamaRes.value.data.data : {};
+
+    return {
+      cmcPrices,
+      llamaPrices,
+    };
+  }, [serverEndpoint]);
+
+  const fetchTokenPrice = useCallback(
+    async ({
+      chainId,
+      chainType,
+      tokenAddress,
+    }: {
+      chainId: number;
+      chainType: ChainType;
+      tokenAddress: string;
+    }) => {
+      const { data } = await axios.get<{ data: number }>(`${serverEndpoint}/api/token/v2/price`, {
+        params: {
+          chainId,
+          tokenAddress: isNativeToken(tokenAddress, chainType) ? undefined : tokenAddress,
+        },
+      });
+      return data;
+    },
+    [serverEndpoint],
+  );
+
   return {
     getTokenPrice,
+    fetchTokenPrice,
+    fetchApiTokenPrices,
   };
 }