+++ packages/canonical-bridge-widget/src/modules/transfer/components/ToTokenSection/index.tsx
@@ -9,7 +9,7 @@ import {
   Link,
   Center,
 } from '@bnb-chain/space';
-import { IBridgeToken, isSameAddress } from '@bnb-chain/canonical-bridge-sdk';
+import { IBridgeToken, isNativeToken, isSameAddress } from '@bnb-chain/canonical-bridge-sdk';
 
 import { useAppSelector } from '@/modules/store/StoreProvider';
 import { IconImage } from '@/core/components/IconImage';
@@ -66,19 +66,19 @@ function ToTokenItem({ token, isSelected }: { token: IBridgeToken; isSelected: b
   const { selectToToken } = useSelection();
   const { isMobile } = useResponsive();
 
-  const fromChain = useAppSelector((state) => state.transfer.fromChain);
+  const toChain = useAppSelector((state) => state.transfer.toChain);
   const isGlobalFeeLoading = useAppSelector((state) => state.transfer.isGlobalFeeLoading);
 
-  const tokenUrl = formatTokenUrl(fromChain?.tokenUrlPattern, token.address);
+  const tokenUrl = formatTokenUrl(toChain?.tokenUrlPattern, token.address);
 
   return (
     <InfoTooltip label={token.name}>
       <Flex
         className="bccb-widget-to-token-item"
-        data-to-name={token.name}
-        data-to-address={token.address}
-        data-to-symbol={token.symbol}
-        data-to-display-symbol={token.displaySymbol}
+        data-available-to-name={token.name}
+        data-available-to-address={token.address}
+        data-available-to-symbol={token.symbol}
+        data-available-to-display-symbol={token.displaySymbol}
         p={{ base: '8px', md: '8px 12px' }}
         h="48px"
         alignItems="center"
@@ -129,26 +129,39 @@ function ToTokenItem({ token, isSelected }: { token: IBridgeToken; isSelected: b
             {token.displaySymbol}
           </Text>
 
-          {isMobile ? (
-            <Flex alignItems="center" gap="4px" textDecoration={{ base: 'underline', md: 'unset' }}>
-              {formatAppAddress({
-                address: token.address,
-              })}
-              <Center boxSize="16px">
-                {isSelected && (
-                  <Link isExternal href={tokenUrl} color="inherit" _hover={{ color: 'inherit' }}>
-                    <ExLinkIcon boxSize="100%" />
-                  </Link>
-                )}
-              </Center>
-            </Flex>
-          ) : (
-            <Link isExternal href={tokenUrl} color="inherit" _hover={{ color: 'inherit' }}>
-              {formatAppAddress({
-                address: token.address,
-                headLen: 4,
-              })}
-            </Link>
+          {!isNativeToken(token.address, toChain?.chainType) && (
+            <>
+              {isMobile ? (
+                <Flex
+                  alignItems="center"
+                  gap="4px"
+                  textDecoration={{ base: 'underline', md: 'unset' }}
+                >
+                  {formatAppAddress({
+                    address: token.address,
+                  })}
+                  <Center boxSize="16px">
+                    {isSelected && (
+                      <Link
+                        isExternal
+                        href={tokenUrl}
+                        color="inherit"
+                        _hover={{ color: 'inherit' }}
+                      >
+                        <ExLinkIcon boxSize="100%" />
+                      </Link>
+                    )}
+                  </Center>
+                </Flex>
+              ) : (
+                <Link isExternal href={tokenUrl} color="inherit" _hover={{ color: 'inherit' }}>
+                  {formatAppAddress({
+                    address: token.address,
+                    headLen: 4,
+                  })}
+                </Link>
+              )}
+            </>
           )}
         </Flex>
       </Flex>